Implementing Different Concepts of Prototype.
here we first simply make a object and intitilse with some valuse.
here we use 'this' which referred to window object and then we use call method for current object.
here we use 'new' keyword which refer to anonymous object which is assign to object and finally define method in object in the class,which can cuase headache so we define this method outside the and to prevent Global pollution we make an obect which define all the sharabel methods which is linked by:
this.__proto__=methos_name.methods.

  This all about first part of Javascript prototype
  ---->> wait for next part.